bayesian_backend/pyproject.toml
[project]
name = "bayesian-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matlabengine>=24.2.1",
    "matplotlib>=3.9.2",
    "optuna>=4.0.0",
    "plotly>=5.24.1",
    "polars>=1.8.2",
    "psycopg2>=2.9.9",
    "python-dotenv>=1.0.1",
    "ruff>=0.6.8",
]
